requirements.txt
flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3
spacy==3.7.2
regex==2023.12.25
